# internal/httpapi/router.go
package httpapi

import (
	"net/http"

	"github.com/utkayd/qurator/internal/observability"
)

// Middleware is the standard wrapping shape.
type Middleware func(http.Handler) http.Handler

// Handlers are the per-stream handler sets. Each Stage 2 stream fills its own; the
// foundation supplies stubs. Nil entries fall back to notImplemented.
type Handlers struct {
	Public    http.Handler // /r/{code}, /i/{id}.{ext}
	QR        http.Handler // /v1/qr
	Codes     http.Handler // /v1/codes*
	Auth      http.Handler // /v1/auth/*
	Tokens    http.Handler // /v1/tokens*
	Admin     http.Handler // /v1/admin/*
	Analytics http.Handler // /v1/codes/{id}/analytics
	Export    http.Handler // /v1/export
	Console   http.Handler // /ui/*
	Healthz   http.Handler
	Readyz    http.Handler
}

// Options configures the two route groups.
//
// Common middleware runs on EVERY request (request ID, metrics, recovery, logging).
// Auth runs ONLY on the protected group. The public group never sees it — this is
// Constitution Principle IV made structural, and router_test.go asserts it.
type Options struct {
	Common []Middleware
	// PerRoute wraps each registered handler individually, so middleware that reads
	// r.Pattern (metrics, logging) sees the LEAF pattern ("GET /r/{code}") rather than
	// the group prefix ("/v1/") that a nested ServeMux reports to outer middleware.
	PerRoute []Middleware
	Auth     Middleware
	CSRF     Middleware // cookie-session CSRF check; protected group only
	// SigninLimiter wraps both password entry points with the same peer quota.
	// Other console routes and public scan/image routes never consume that quota.
	SigninLimiter Middleware
}

// SigninPattern is the public API password route. Console sign-in shares its limiter.
const SigninPattern = "POST /v1/auth/signin"

// Routes is the full route table from contracts/openapi.yaml plus the console.
// Every entry is registered even when its handler is a stub, so Stage 2 streams fill
// handlers rather than editing this file.
var Routes = []Route{
	// Public — NO auth middleware, ever.
	{Group: GroupPublic, Pattern: "GET /r/{code}", Handler: "Public"},
	{Group: GroupPublic, Pattern: "GET /i/{file}", Handler: "Public"},
	{Group: GroupPublic, Pattern: "GET /healthz", Handler: "Healthz"},
	{Group: GroupPublic, Pattern: "GET /readyz", Handler: "Readyz"},
	{Group: GroupPublic, Pattern: SigninPattern, Handler: "Auth"},

	// Protected — full chain.
	{Group: GroupProtected, Pattern: "GET /v1/qr", Handler: "QR"},
	{Group: GroupProtected, Pattern: "POST /v1/qr", Handler: "QR"},
	{Group: GroupProtected, Pattern: "GET /v1/codes", Handler: "Codes"},
	{Group: GroupProtected, Pattern: "POST /v1/codes", Handler: "Codes"},
	{Group: GroupProtected, Pattern: "POST /v1/codes/batch", Handler: "Codes"},
	{Group: GroupProtected, Pattern: "GET /v1/codes/{id}", Handler: "Codes"},
	{Group: GroupProtected, Pattern: "PATCH /v1/codes/{id}", Handler: "Codes"},
	{Group: GroupProtected, Pattern: "DELETE /v1/codes/{id}", Handler: "Codes"},
	{Group: GroupProtected, Pattern: "POST /v1/codes/{id}/disable", Handler: "Codes"},
	{Group: GroupProtected, Pattern: "POST /v1/codes/{id}/enable", Handler: "Codes"},
	{Group: GroupProtected, Pattern: "GET /v1/codes/{id}/analytics", Handler: "Analytics"},
	{Group: GroupProtected, Pattern: "POST /v1/auth/signout", Handler: "Auth"},
	{Group: GroupProtected, Pattern: "GET /v1/auth/me", Handler: "Auth"},
	{Group: GroupProtected, Pattern: "GET /v1/tokens", Handler: "Tokens"},
	{Group: GroupProtected, Pattern: "POST /v1/tokens", Handler: "Tokens"},
	{Group: GroupProtected, Pattern: "DELETE /v1/tokens/{id}", Handler: "Tokens"},
	{Group: GroupProtected, Pattern: "DELETE /v1/admin/aliases/{alias}", Handler: "Admin"},
	{Group: GroupProtected, Pattern: "GET /v1/export", Handler: "Export"},

	// Console — protected group; the console handler renders its own sign-in page for
	// anonymous requests, so /ui/ is mounted with auth in "optional" mode by the stream.
	{Group: GroupConsole, Pattern: "/ui/", Handler: "Console"},
}

// Group names a route group.
type Group string

const (
	GroupPublic    Group = "public"
	GroupProtected Group = "protected"
	GroupConsole   Group = "console"
)

// Route is one registered pattern.
type Route struct {
	Group   Group
	Pattern string
	Handler string // field name in Handlers
}

// NewRouter builds the root handler. The ephemeral endpoint (/v1/qr) sits in the
// protected group; when config makes it public the QR handler itself decides to accept
// anonymous identities — the auth middleware attaches identity but does not reject on
// that route (see auth stream). Public scan routes are a separate mux that the auth
// middleware is never applied to.
func NewRouter(h Handlers, o Options) http.Handler {
	public := http.NewServeMux()
	protected := http.NewServeMux()
	console := http.NewServeMux()

	for _, rt := range Routes {
		hd := withPattern(rt.Pattern, h.lookup(rt.Handler))
		if rt.Pattern == SigninPattern && o.SigninLimiter != nil {
			hd = o.SigninLimiter(hd)
		}
		if rt.Group == GroupConsole && o.SigninLimiter != nil {
			next := hd
			limited := o.SigninLimiter(next)
			hd = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
				if r.Method == http.MethodPost && r.URL.Path == "/ui/signin" {
					limited.ServeHTTP(w, r)
					return
				}
				next.ServeHTTP(w, r)
			})
		}
		for i := len(o.PerRoute) - 1; i >= 0; i-- {
			hd = o.PerRoute[i](hd)
		}
		switch rt.Group {
		case GroupPublic:
			public.Handle(rt.Pattern, hd)
		case GroupProtected:
			protected.Handle(rt.Pattern, hd)
		case GroupConsole:
			console.Handle(rt.Pattern, hd)
		}
	}

	protectedChain := jsonNotFound(protected)
	if o.CSRF != nil {
		protectedChain = o.CSRF(protectedChain)
	}
	if o.Auth != nil {
		protectedChain = o.Auth(protectedChain)
	}
	var consoleChain http.Handler = console
	if o.Auth != nil {
		consoleChain = o.Auth(consoleChain)
	}

	root := http.NewServeMux()
	root.Handle("/r/", public)
	root.Handle("/i/", public)
	root.Handle("/healthz", public)
	root.Handle("/readyz", public)
	root.Handle(SigninPattern, public)
	root.Handle("/v1/", protectedChain)
	root.Handle("/ui/", consoleChain)
	root.Handle("/", http.RedirectHandler("/ui/", http.StatusFound))

	var out http.Handler = root
	for i := len(o.Common) - 1; i >= 0; i-- {
		out = o.Common[i](out)
	}
	return out
}

func (h Handlers) lookup(name string) http.Handler {
	var hd http.Handler
	switch name {
	case "Public":
		hd = h.Public
	case "QR":
		hd = h.QR
	case "Codes":
		hd = h.Codes
	case "Auth":
		hd = h.Auth
	case "Tokens":
		hd = h.Tokens
	case "Admin":
		hd = h.Admin
	case "Analytics":
		hd = h.Analytics
	case "Export":
		hd = h.Export
	case "Console":
		hd = h.Console
	case "Healthz":
		hd = h.Healthz
	case "Readyz":
		hd = h.Readyz
	}
	if hd == nil {
		return NotImplemented()
	}
	return hd
}

// NotImplemented is the foundation stub for every route a stream has not filled yet.
func NotImplemented() http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		WriteError(w, CodeNotImplemented, "This endpoint is not implemented yet.", map[string]any{"route": r.Pattern})
	})
}

// jsonNotFound wraps a mux so an unmatched path answers with the JSON error envelope
// (FR-044) instead of the ServeMux's plain-text "404 page not found". A matched path
// with the wrong method is left to the mux, which answers 405 with an Allow header.
func jsonNotFound(mux *http.ServeMux) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		if _, pattern := mux.Handler(r); pattern != "" {
			mux.ServeHTTP(w, r)
			return
		}
		// Unmatched: ask the mux whether it would say 404 or 405 without letting it
		// write to the client.
		probe := &statusProbe{header: http.Header{}}
		mux.ServeHTTP(probe, r)
		if probe.code != http.StatusNotFound {
			mux.ServeHTTP(w, r)
			return
		}
		WriteError(w, CodeNotFound, "No such endpoint.", nil)
	})
}

// statusProbe records only the status code of a response and discards the body.
type statusProbe struct {
	header http.Header
	code   int
}

func (p *statusProbe) Header() http.Header       { return p.header }
func (p *statusProbe) WriteHeader(code int)      { p.code = code }
func (p *statusProbe) Write([]byte) (int, error) { return 0, nil }

// withPattern stamps the registered pattern into the request context so
// observability.RoutePattern is authoritative regardless of nested muxes.
func withPattern(pattern string, next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		next.ServeHTTP(w, r.WithContext(observability.WithRoutePattern(r.Context(), pattern)))
	})
}