# specs/001-qr-service-baseline/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: qurator API
  version: "1.0.0"
  summary: Self-hostable QR code service — ephemeral generation, dynamic codes, redirects, analytics, tokens.
  description: |
    qurator is a single self-hostable service exposing:

    - **Ephemeral generation** (`/v1/qr`): render a QR image from content and styling with no
      storage side effect (FR-001..FR-006).
    - **Dynamic codes** (`/v1/codes`): persisted, mutable-destination codes with a stable
      printed image (FR-007..FR-018).
    - **Public redirect and image serving** (`/r/{code}`, `/i/{id}.{ext}`): unauthenticated,
      latency-sensitive, and structurally isolated from the analytics and metadata write path
      (FR-009, FR-015, FR-017).
    - **Analytics** (`/v1/codes/{id}/analytics`): totals, a time series, and per-dimension
      breakdowns — deliberately with **no geographic dimension** and **no stored scanner IP**
      in v1 (FR-022, FR-025).
    - **Identity and tokens** (`/v1/auth/*`, `/v1/tokens`): one credential-verification path
      for both browser sessions and machine tokens (FR-031..FR-038).
    - **Export** (`/v1/export`): a streaming, machine-readable dump of everything an operator
      needs to walk away with (FR-055).

    ### Authentication

    Two schemes are accepted **interchangeably** on every protected operation: a bearer API
    token (`Authorization: Bearer qur_<secret>`) or the HttpOnly session cookie set at sign-in.
...
    absence.

...

    ### CSRF

...

    ### Errors

...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /r/{code}:
    get:
      operationId: redirectScan
      tags: [public]
      summary: Resolve a short code and redirect the scanner
      description: |
        Resolves `code` to its current destination and issues a redirect (FR-009). This path
        carries no authentication middleware whatsoever (Principle IV) and performs at most
        one metadata lookup, fronted by an in-process cache with a bounded staleness window
        (FR-017).

        **Why `302` and not `301`/`308`.** A permanent redirect status is heuristically
        cached indefinitely by browsers and intermediaries. Because a dynamic code's whole
        purpose is that its destination can change after printing (US2), a cached permanent
        redirect would silently keep sending scanners to a stale destination and would stop
        the scan from ever reaching the instance to be counted. `302` is paired with
        `Cache-Control: no-store, no-cache, must-revalidate` so every scan reaches the
        instance, keeping destination changes effective within seconds (SC-008) and analytics
        accurate (FR-020).

        An unknown, disabled, or deleted code does **not** error — it returns a clear,
        non-error-looking landing response rather than a redirect (FR-014), so a scan of a
        retired code never looks like a broken page and is never expensive enough to serve as
        an enumeration vector.
      security: []
      parameters:
        - name: code
          in: path
          required: true
          description: The short code or custom alias, matched case-insensitively.
          schema:
            $ref: '#/components/schemas/ShortCode'
      responses:
        '302':
          description: |
            The code is active. The scanner is redirected to its current destination.
            Always non-cacheable so the next scan re-resolves the destination.
          headers:
            Location:
              description: The code's current destination URL.
              schema:
                type: string
                format: uri
              example: https://example.com/spring-sale
            Cache-Control:
              description: Prevents any intermediary from caching this redirect.
              schema:
                type: string
                enum: ["no-store, no-cache, must-revalidate"]
              example: "no-store, no-cache, must-revalidate"
        '200':
          description: |
            The code is unknown, disabled, or deleted. A human-readable landing page is
            returned (FR-014) — never a stack trace, never a generic server error, and never
            a silent redirect to an unexpected destination. If the operator has configured a
            fallback destination for this case, the landing response says so.
          content:
            text/html:
              schema:
                type: string
        '5XX':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/qr:
    get:
      operationId: generateQrGet
      tags: [generation]
      summary: Render an ephemeral QR image (query-string form)
      description: |
        Renders content directly to image bytes with **no metadata-store or blob-store
        record created** (FR-001, FR-003). Identical input produces byte-identical output, so
        responses are safely cacheable by intermediaries (FR-004).

        **Authentication**: required by default (FR-006). An operator may explicitly set
        `generation.allow_unauthenticated: true`, in which case this operation accepts
        anonymous requests subject to a configurable rate limit — the two modes are described
        together here because the same operation behaves differently under each, not because
        both apply at once on a given instance.

        Use `GET` for simple content that fits in a URL; use `POST` (below) when content or
        styling would make the query string unwieldy or when content contains characters
        awkward to percent-encode.
      security:
        - bearerAuth: []
        - cookieAuth: []
        - {}
      parameters:
        - name: content
          in: query
          required: true
          description: The payload to encode. Maximum length depends on the effective EC level (2953 bytes at L, down to 1273 at H).
          schema:
            type: string
          example: "https://example.com/hello"
        - name: format
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ImageFormat'
        - name: fg_color
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/HexColor'
        - name: bg_color
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/HexColor'
        - name: module_shape
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ModuleShape'
        - name: margin_modules
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/MarginModules'
        - name: size_px
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/SizePx'
        - name: ec_level
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/EcLevel'
      responses:
        '200':
          $ref: '#/components/responses/RenderedImage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/ContentTooLarge'
        '429':
          $ref: '#/components/responses/RateLimited'
        '5XX':
          description: Unexpected server error, or a render that exceeded the maximum rendering duration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      operationId: generateQrPost
      tags: [generation]
      summary: Render an ephemeral QR image (JSON body form)
      description: |
        Identical semantics to `GET /v1/qr`, accepting content and styling as a JSON body.
        Preferred for binary or large content, or a full styling object including a logo.
        See `GET /v1/qr` for the authentication and no-storage-record guarantees, which apply
        identically here.
      security:
        - bearerAuth: []
        - cookieAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateQrRequest'
            example:
              content: "https://example.com/hello"
              format: png
              styling:
                fg_color: "#101828"
                bg_color: "#FFFFFF"
                module_shape: dot
                margin_modules: 4
                size_px: 512
                ec_level: M
      responses:
        '200':
          $ref: '#/components/responses/RenderedImage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/ContentTooLarge'
        '429':
          $ref: '#/components/responses/RateLimited'
        '5XX':
          description: Unexpected server error, or a render that exceeded the maximum rendering duration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---------------------------------------------------------------------
  # Dynamic codes
  # ---------------------------------------------------------------------
  /v1/codes:
    get:
      operationId: listCodes
      tags: [codes]
      summary: List the caller's dynamic codes
      description: Paginated, and filterable by creation time and by destination (FR-016).
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          example: "eyJpZCI6ImNvZF9hYmMxMjMifQ"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: created_after
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: created_before
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: destination_contains
          in: query
          required: false
          description: Case-insensitive substring filter over the current destination URL.
          schema:
            type: string
          example: "example.com"
      responses:
        '200':
          description: A page of the caller's codes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodePage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '5XX':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      operationId: createCode
      tags: [codes]
      summary: Create a dynamic or direct code
      description: |
        Creates a persisted code with a mutable destination (FR-007). If `alias` is omitted, a
        system-generated, unguessable, collision-resistant short code is assigned (FR-013). If
        `alias` is supplied, it is validated per FR-018: unique case-insensitively, restricted
        to a safe URL character set, not on the reserved-word list, and not previously used by
        a deleted code (unless an administrator has explicitly released it). Aliases are
        **immutable** once created.

        The response's `image_url` addresses the persisted image according to `images.url_mode`.
        Dynamic codes encode the configured absolute `server.base_url` plus `/r/{short_code}`;
        direct codes encode the destination itself (spec 002).
        New dynamic creation requires a configured scan origin; an empty origin returns
        `503 scan_url_not_configured` before rendering or persistence (spec 004).
        Direct creation and idempotent replays of existing codes remain available.
        Request Host and forwarded headers are never used as the printed origin.
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCodeRequest'
            example:
              destination: "https://example.com/spring-sale"
              alias: "spring-sale"
              styling:
                fg_color: "#101828"
                bg_color: "#FFFFFF"
                module_shape: rounded
                margin_modules: 4
                size_px: 1024
                ec_level: Q
      responses:
        '200':
          description: Existing code returned for an idempotent client_ref replay.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Code'
        '201':
          description: The code was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Code'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          description: Dynamic creation is unavailable until the operator configures server.base_url.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: scan_url_not_configured
                  message: Dynamic codes need a public scan address. Ask the operator to configure QURATOR_SERVER_BASE_URL, or create a direct code.
        '5XX':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/codes/{id}:
    get:
      operationId: getCode
      tags: [codes]
      summary: Get a single dynamic code
      parameters:
        - $ref: '#/components/parameters/CodeIdPath'
      responses:
        '200':
          description: The code.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Code'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '5XX':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      operationId: updateCodeDestination
      tags: [codes]
      summary: Update a dynamic code's destination
      description: |
        Changes only the destination (FR-008). The encoded QR value and the stored image are
        never altered by this operation (FR-010) — restyling is not available on an existing
        code; a differently styled printed artefact is a new code (spec Assumptions:
        "Image immutability").

        The new destination is validated against the configured scheme allow-list and against
        self-referential resolution (loop back to this instance's own `/r/` path), exactly as
        on creation (FR-011, FR-012). On rejection, the existing destination is left unchanged.

        Concurrent edits are resolved deterministically by an optimistic-concurrency check:
        pass the code's current `version` as a quoted entity-tag in `If-Match`
        (e.g. `If-Match: "7"`), or accept last-write-wins semantics by omitting it. A losing
        conditional write is refused with `409` and the current version in `details`, never
        silently discarded (Edge Cases: "Concurrent destination edits").

        `If-Match` on an integer version is used rather than `If-Unmodified-Since` on
        `updated_at` because HTTP-date has one-second granularity: two edits within the same
        second would both pass a timestamp check, which is precisely the race this exists to
        prevent.
      parameters:
        - $ref: '#/components/parameters/CodeIdPath'
        - $ref: '#/components/parameters/CsrfHeader'
        - name: If-Match
          in: header
          required: false
          description: >-
            The code's `version` last observed by the caller, as a quoted entity-tag
            (e.g. `"7"`). Omit to accept last-write-wins.
          schema:
            type: string
            pattern: '^"[0-9]+"$'
            example: '"7"'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDestinationRequest'
            example:
              destination: "https://example.com/spring-sale-2026"
      responses:
        '200':
          description: The updated code.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Code'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '5XX':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      operationId: deleteCode
      tags: [codes]
      summary: Soft-delete a dynamic code
      description: |
        Marks the code `deleted` (terminal). Its short code / alias remains reserved and
        cannot be reassigned to a different code (FR-018), so a printed artefact can never be
        hijacked by a later registration. A subsequent scan returns the disabled/deleted
        landing response, not a redirect.
      parameters:
        - $ref: '#/components/parameters/CodeIdPath'
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '204':
          description: The code was deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '5XX':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/codes/{id}/disable:
    post:
      operationId: disableCode
      tags: [codes]
      summary: Disable a dynamic code
      description: |
        Transitions the code from `active` to `disabled`. A disabled code's scans receive the
        non-error landing response rather than a redirect (FR-014); the code can later be
        re-enabled.
      parameters:
        - $ref: '#/components/parameters/CodeIdPath'
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '200':
          description: The code, now disabled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Code'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The code is `deleted`, which is terminal and cannot transition back to `disabled`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '5XX':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/codes/{id}/enable:
    post:
      operationId: enableCode
      tags: [codes]
      summary: Re-enable a disabled dynamic code
      parameters:
        - $ref: '#/components/parameters/CodeIdPath'
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '200':
          description: The code, now active.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Code'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The code is `deleted`, which is terminal and cannot transition back to `active`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '5XX':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---------------------------------------------------------------------
  # Analytics
  # ---------------------------------------------------------------------
  /v1/codes/batch:
    post:
      operationId: createCodesBatch
      tags: [codes]
      summary: Create up to `codes.batch_max` dynamic or direct codes in one request
      description: |
        Each item is an independent `CreateCodeRequest`. The response is per item: a
        created code, the existing code when the item's `client_ref` was already used
        with the same destination and mode (`status: existing`), or an error envelope
        for that item alone. The HTTP status is `200` whenever the batch was processed;
        only a malformed or oversized batch fails as a whole. With an empty scan origin,
        new dynamic items report `scan_url_not_configured`; direct and existing items
        can still succeed (spec 004). Metadata for all created
        items is written in one transaction; images render in parallel (spec 003).
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              additionalProperties: false
              properties:
                items:
                  type: array
                  minItems: 1
                  items:
                    $ref: '#/components/schemas/CreateCodeRequest'
      responses:
        '200':
          description: Per-item results in input order.
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      required: [index, status]
                      properties:
                        index: { type: integer }
                        status: { type: string, enum: [created, existing, error] }
                        code:
                          $ref: '#/components/schemas/Code'
                        error:
                          $ref: '#/components/schemas/ErrorDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: More items than `codes.batch_max`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '5XX':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/auth/signin:
    post:
      operationId: signIn
      tags: [auth]
      summary: Sign in with local credentials
      description: |
        Verifies email and password against the local account (irrelevant when the caller is
        authenticated exclusively via a forward-auth proxy; see FR-036). On success, sets the
        HttpOnly session cookie. There is no self-service registration or password reset
        (FR-039) — only the bootstrap admin account and any accounts it creates can sign in
        this way.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInRequest'
            example:
              email: admin@example.com
              password: "correct horse battery staple"
      responses:
        '200':
          description: Signed in. The session cookie is set via `Set-Cookie`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '5XX':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/auth/signout:
    post:
      operationId: signOut
      tags: [auth]
      summary: Sign out and clear the session cookie
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '204':
          description: Signed out.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '5XX':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "qur_<base64url secret>"
      description: |
        An API token, presented as `Authorization: Bearer qur_<base64url of 32 random
        bytes>`. Verified through the same identity-resolution path as `cookieAuth` (FR-031).
        Bearer-authenticated requests do NOT need the `X-Qurator-Requested-With` CSRF header,
        since a bearer token is never sent implicitly by a browser.
    cookieAuth:
      type: apiKey
      in: cookie
      name: qurator_session
      description: |
        An HttpOnly, `SameSite=Strict` session cookie set by `POST /v1/auth/signin`,
        restricted to this instance's origin. It is `Secure` unless `server.base_url`
        is an explicit `http://` origin. State-changing operations authenticated this
        way MUST also carry the `X-Qurator-Requested-With` header (see the `CsrfHeader`
        parameter) as a same-origin assertion.
...